added controls for bad request

+++ api/controllers/Items/remove.js
@@ -3,6 +3,9 @@ module.exports = {
   async fn(inputs, exits, env) {
     const {res,req} = env;
     const {id, idList} = req.query;
+    if (!id) {
+      return res.badRequest('id is not being sent');
+    }
     const itemRemoved = await ItemsService.remove({id});
     if (itemRemoved && itemRemoved.error){
       return res.serverError(itemsCreated);
