+++ api/controllers/Shopping-lists/remove.js
@@ -3,6 +3,9 @@ module.exports = {
   async fn(inputs, exits, env) {
     const {res,req} = env;
     const {id, createdBy} = req.query;
+    if (!id) {
+      return res.badRequest('id is not being sent');
+    }
     let list = await ShoppingListsService.remove({id, createdBy});
     if (list && list.error){
       return res.serverError(list);
